requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
segno
python-dotenv